# Set CPUs for `soltest.sh` based on the number of available cores