@cameel 2023-08-07
The name of this variable is a bit misleading. It's not meant to literally represent the number of CPUs. More like the number of processes to run on those cores. The usual recommendation is to run N+1 jobs on N cores to saturate them. The idea is that when a process becomes blocked on an I/O operation, the core is basically free and some other process can take it over. I've even seen 1.5N+1 or 2N+1 recommended, but that does not seem necessary here since the graphs already show 100% usage.

I considered renaming it, but I don't really have a good name. We already have "parallelism" and "jobs" used for CircleCI things so naming gets confusing.

I could make it literally mean the number of CPUs and have the script add +1, but I like that it matches the value passed to `-j`. Makes it easier to see that the value is right.