@blishko 2023-08-09
OK! Thanks for the explanation. From my side, this is OK.
The memory issues I have been experiencing are local to my branch.

Probably someone else should also look at this and approve?