@cameel 2023-08-09
Well, this is just a CI change and it's pretty simple so it does not require a very strict review. I think we can just merge it if you don't see any obvious problems there and approve.